requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
pycryptodome==3.19.0
python-dotenv==1.0.0
